fix: reorganize code
