+++ src/lib.rs
@@ -35,9 +35,14 @@
 //! Alternative storage backends can be implemented by defining a type that
 //! implements the `DimensionalStorage` trait.
 
+mod linear_storage;
+mod operators;
+
+pub use linear_storage::{LinearArrayLayout, LinearArrayStorage};
+
 use num::Num;
 use std::marker::PhantomData;
-use std::ops::{Add, Div, Index, IndexMut, Mul, Sub};
+use std::ops::{Index, IndexMut};
 
 /// A trait for storage backends for multidimensional arrays.
 ///
@@ -72,134 +77,6 @@ pub trait DimensionalStorage<T: Num + Copy, const N: usize>:
     /// * `shape`: The shape of the array.
     /// * `data`: The data to initialize the array with.
     fn from_vec(shape: [usize; N], data: Vec<T>) -> Self;
-}
-
-/// An enum representing the memory layout of a linear array.
-#[derive(Debug, Copy, Clone, PartialEq)]
-enum LinearArrayLayout {
-    /// Row-major layout (default).
-    RowMajor,
-    // TODO: figure out if we want to support column-major layout
-    #[allow(dead_code)]
-    ColumnMajor,
-}
-
-/// A linear array storage backend for multidimensional arrays.
-///
-/// This struct stores the array data in a contiguous block of memory,
-/// using either row-major or column-major layout.
-///
-/// # Type Parameters
-///
-/// * `T`: The element type of the array. Must implement `Num` and `Copy`.
-/// * `N`: The number of dimensions of the array.
-#[derive(Debug, Clone, PartialEq)]
-pub struct LinearArrayStorage<T: Num + Copy, const N: usize> {
-    data: Vec<T>,
-    layout: LinearArrayLayout,
-    strides: [usize; N],
-}
-
-impl<T: Num + Copy, const N: usize> Index<[usize; N]> for LinearArrayStorage<T, N> {
-    type Output = T;
-
-    fn index(&self, index: [usize; N]) -> &Self::Output {
-        let linear_index = self.layout_index(index);
-        &self.data[linear_index]
-    }
-}
-
-impl<T: Num + Copy, const N: usize> IndexMut<[usize; N]> for LinearArrayStorage<T, N> {
-    fn index_mut(&mut self, index: [usize; N]) -> &mut Self::Output {
-        let linear_index = self.layout_index(index);
-        &mut self.data[linear_index]
-    }
-}
-
-impl<T: Num + Copy, const N: usize> DimensionalStorage<T, N> for LinearArrayStorage<T, N> {
-    fn zeros(shape: [usize; N]) -> Self {
-        let data = vec![T::zero(); shape.iter().product::<usize>()];
-        LinearArrayStorage::new(shape, data, LinearArrayLayout::RowMajor, 1)
-    }
-
-    fn ones(shape: [usize; N]) -> Self {
-        let data = vec![T::one(); shape.iter().product::<usize>()];
-        LinearArrayStorage::new(shape, data, LinearArrayLayout::RowMajor, 1)
-    }
-
-    fn from_vec(shape: [usize; N], data: Vec<T>) -> Self {
-        LinearArrayStorage::new(shape, data, LinearArrayLayout::RowMajor, 1)
-    }
-}
-
-impl<T: Num + Copy, const N: usize> LinearArrayStorage<T, N> {
-    /// Computes the strides for a given shape and layout.
-    ///
-    /// # Arguments
-    ///
-    /// * `shape`: The shape of the array.
-    /// * `stride`: The base stride (usually 1).
-    /// * `layout`: The memory layout of the array.
-    fn compute_strides(
-        shape: &[usize; N],
-        stride: &usize,
-        layout: &LinearArrayLayout,
-    ) -> [usize; N] {
-        let mut strides = [0; N];
-        match layout {
-            LinearArrayLayout::RowMajor => {
-                strides[N - 1] = *stride;
-                for i in (0..N - 1).rev() {
-                    strides[i] = strides[i + 1] * shape[i + 1];
-                }
-            }
-            LinearArrayLayout::ColumnMajor => {
-                strides[0] = *stride;
-                for i in 1..N {
-                    strides[i] = strides[i - 1] * shape[i - 1];
-                }
-            }
-        }
-        strides
-    }
-
-    /// Computes the linear index for a given multidimensional index.
-    ///
-    /// # Arguments
-    ///
-    /// * `index`: The multidimensional index.
-    fn layout_index(&self, index: [usize; N]) -> usize {
-        match self.layout {
-            LinearArrayLayout::RowMajor => index
-                .iter()
-                .zip(self.strides.iter())
-                .map(|(i, &stride)| i * stride)
-                .sum(),
-            LinearArrayLayout::ColumnMajor => index
-                .iter()
-                .rev()
-                .zip(self.strides.iter().rev())
-                .map(|(i, &stride)| i * stride)
-                .sum(),
-        }
-    }
-
-    /// Creates a new `LinearArrayStorage` with the given parameters.
-    ///
-    /// # Arguments
-    ///
-    /// * `shape`: The shape of the array.
-    /// * `data`: The data to initialize the array with.
-    /// * `layout`: The memory layout of the array.
-    /// * `stride`: The base stride (usually 1).
-    fn new(shape: [usize; N], data: Vec<T>, layout: LinearArrayLayout, stride: usize) -> Self {
-        let strides = Self::compute_strides(&shape, &stride, &layout);
-        Self {
-            data,
-            layout,
-            strides,
-        }
-    }
 }
 
 /// A multidimensional array type.
@@ -347,156 +224,6 @@ where
     }
 }
 
-impl<T: Num + Copy, S, const N: usize> Index<[usize; N]> for Dimensional<T, S, N>
-where
-    S: DimensionalStorage<T, N>,
-{
-    type Output = T;
-
-    fn index(&self, index: [usize; N]) -> &Self::Output {
-        &self.storage[index]
-    }
-}
-
-impl<T: Num + Copy, S, const N: usize> IndexMut<[usize; N]> for Dimensional<T, S, N>
-where
-    S: DimensionalStorage<T, N>,
-{
-    fn index_mut(&mut self, index: [usize; N]) -> &mut Self::Output {
-        &mut self.storage[index]
-    }
-}
-
-// Generalized operations
-
-impl<T: Num + Copy + PartialEq, S, const N: usize> PartialEq for Dimensional<T, S, N>
-where
-    S: DimensionalStorage<T, N>,
-{
-    fn eq(&self, other: &Self) -> bool {
-        if self.shape != other.shape {
-            return false;
-        }
-
-        for i in 0..self.shape.iter().product::<usize>() {
-            let index = Self::unravel_index(i, &self.shape);
-            if self[index] != other[index] {
-                return false;
-            }
-        }
-
-        true
-    }
-}
-
-// Scalar-vector operations
-impl<T: Num + Copy, S: DimensionalStorage<T, 1>> Add<T> for &Dimensional<T, S, 1> {
-    type Output = Dimensional<T, S, 1>;
-
-    fn add(self, rhs: T) -> Self::Output {
-        let mut result = Dimensional::zeros(self.shape);
-        for i in 0..self.shape[0] {
-            result[[i]] = self[[i]] + rhs;
-        }
-        result
-    }
-}
-
-impl<T: Num + Copy, S: DimensionalStorage<T, 1>> Sub<T> for &Dimensional<T, S, 1> {
-    type Output = Dimensional<T, S, 1>;
-
-    fn sub(self, rhs: T) -> Self::Output {
-        let mut result = Dimensional::zeros(self.shape);
-        for i in 0..self.shape[0] {
-            result[[i]] = self[[i]] - rhs;
-        }
-        result
-    }
-}
-
-impl<T: Num + Copy, S: DimensionalStorage<T, 1>> Mul<T> for &Dimensional<T, S, 1> {
-    type Output = Dimensional<T, S, 1>;
-
-    fn mul(self, rhs: T) -> Self::Output {
-        let mut result = Dimensional::zeros(self.shape);
-        for i in 0..self.shape[0] {
-            result[[i]] = self[[i]] * rhs;
-        }
-        result
-    }
-}
-
-impl<T: Num + Copy, S: DimensionalStorage<T, 1>> Div<T> for &Dimensional<T, S, 1> {
-    type Output = Dimensional<T, S, 1>;
-
-    fn div(self, rhs: T) -> Self::Output {
-        let mut result = Dimensional::zeros(self.shape);
-        for i in 0..self.shape[0] {
-            result[[i]] = self[[i]] / rhs;
-        }
-        result
-    }
-}
-
-// Vector-vector operations
-impl<T: Num + Copy, S: DimensionalStorage<T, 1>> Add<&Dimensional<T, S, 1>>
-    for &Dimensional<T, S, 1>
-{
-    type Output = Dimensional<T, S, 1>;
-
-    fn add(self, rhs: &Dimensional<T, S, 1>) -> Self::Output {
-        assert_eq!(
-            self.shape, rhs.shape,
-            "Cannot add vectors of different lengths"
-        );
-        let mut result = Dimensional::zeros(self.shape);
-        for i in 0..self.shape[0] {
-            result[[i]] = self[[i]] + rhs[[i]];
-        }
-        result
-    }
-}
-
-impl<T: Num + Copy, S: DimensionalStorage<T, 1>> Sub<&Dimensional<T, S, 1>>
-    for &Dimensional<T, S, 1>
-{
-    type Output = Dimensional<T, S, 1>;
-
-    fn sub(self, rhs: &Dimensional<T, S, 1>) -> Self::Output {
-        assert_eq!(
-            self.shape, rhs.shape,
-            "Cannot subtract vectors of different lengths"
-        );
-        let mut result = Dimensional::zeros(self.shape);
-        for i in 0..self.shape[0] {
-            result[[i]] = self[[i]] - rhs[[i]];
-        }
-        result
-    }
-}
-
-// Matrix Math
-
-impl<T: Num + Copy, S: DimensionalStorage<T, 2>> Add<&Dimensional<T, S, 2>>
-    for &Dimensional<T, S, 2>
-{
-    type Output = Dimensional<T, S, 2>;
-
-    // Vector addition
-    fn add(self, rhs: &Dimensional<T, S, 2>) -> Self::Output {
-        let shape = self.shape;
-        let mut result = Dimensional::zeros(shape);
-
-        for i in 0..shape[0] {
-            for j in 0..shape[1] {
-                result[[i, j]] = self[[i, j]] + rhs[[i, j]];
-            }
-        }
-
-        result
-    }
-}
-
 // Macros
 
 #[macro_export]
@@ -536,78 +263,8 @@ macro_rules! matrix {
 mod tests {
 
     use super::*;
-    use crate::{matrix, scalar, vector};
-
-    #[test]
-    fn test_dimensional_array_column_major_layout() {
-        let shape = [2, 3];
-        let data = vec![1.0, 4.0, 2.0, 5.0, 3.0, 6.0];
-        let array = Dimensional::new(
-            shape,
-            LinearArrayStorage::new(shape, data, LinearArrayLayout::ColumnMajor, 1),
-        );
-
-        assert_eq!(array[[0, 0]], 1.0);
-        assert_eq!(array[[1, 0]], 4.0);
-        assert_eq!(array[[0, 1]], 2.0);
-        assert_eq!(array[[1, 1]], 5.0);
-        assert_eq!(array[[0, 2]], 3.0);
-        assert_eq!(array[[1, 2]], 6.0);
-    }
-
-    #[test]
-    fn test_scalar() {
-        let s = scalar!(42);
-        assert_eq!(s.shape(), [1]);
-        assert_eq!(s[[0]], 42);
-    }
-
-    #[test]
-    fn test_vector() {
-        let v = vector![1, 2, 3, 4, 5];
-        assert_eq!(v.shape(), [5]);
-        assert_eq!(v[[0]], 1);
-        assert_eq!(v[[2]], 3);
-        assert_eq!(v[[4]], 5);
-    }
-
-    #[test]
-    fn test_matrix() {
-        let m = matrix![[1, 2, 3], [4, 5, 6], [7, 8, 9]];
-        assert_eq!(m.shape(), [3, 3]);
-        assert_eq!(m[[0, 0]], 1);
-        assert_eq!(m[[1, 1]], 5);
-        assert_eq!(m[[2, 2]], 9);
-    }
-
-    #[test]
-    fn test_zeros() {
-        let z = Dimensional::<f64, LinearArrayStorage<f64, 2>, 2>::zeros([3, 4]);
-        assert_eq!(z.shape(), [3, 4]);
-        assert_eq!(z[[0, 0]], 0.0);
-        assert_eq!(z[[1, 2]], 0.0);
-        assert_eq!(z[[2, 3]], 0.0);
-    }
-
-    #[test]
-    fn test_ones() {
-        let o = Dimensional::<i32, LinearArrayStorage<i32, 3>, 3>::ones([2, 3, 4]);
-        assert_eq!(o.shape(), [2, 3, 4]);
-        assert_eq!(o[[0, 0, 0]], 1);
-        assert_eq!(o[[1, 1, 1]], 1);
-        assert_eq!(o[[1, 2, 3]], 1);
-    }
-
-    #[test]
-    fn test_from_fn() {
-        let f = Dimensional::<f64, LinearArrayStorage<f64, 2>, 2>::from_fn([3, 3], |[i, j]| {
-            (i + j) as f64
-        });
-        assert_eq!(f.shape(), [3, 3]);
-        assert_eq!(f[[0, 0]], 0.0);
-        assert_eq!(f[[1, 1]], 2.0);
-        assert_eq!(f[[2, 2]], 4.0);
-    }
+    use crate::linear_storage::LinearArrayStorage;
+    use crate::{matrix, vector};
 
     #[test]
     fn test_indexing() {
@@ -673,78 +330,5 @@ mod tests {
         let m = matrix![[1, 2, 3], [4, 5, 6], [7, 8, 9]];
         assert_eq!(m.len_axis(0), 3);
         assert_eq!(m.len_axis(1), 3);
-    }
-
-    #[test]
-    fn test_addition() {
-        let v1 = vector![1, 2, 3, 4, 5];
-        let v2 = vector![6, 7, 8, 9, 10];
-        let v3 = &v1 + &v2;
-        assert_eq!(v3[[0]], 7);
-        assert_eq!(v3[[2]], 11);
-        assert_eq!(v3[[4]], 15);
-    }
-
-    #[test]
-    fn test_column_major_addition() {
-        let shape = [2, 3];
-        let data1 = vec![1.0, 4.0, 2.0, 5.0, 3.0, 6.0];
-        let data2 = vec![7.0, 10.0, 8.0, 11.0, 9.0, 12.0];
-        let array1 = Dimensional::new(
-            shape,
-            LinearArrayStorage::new(shape, data1, LinearArrayLayout::ColumnMajor, 1),
-        );
-        let array2 = Dimensional::new(
-            shape,
-            LinearArrayStorage::new(shape, data2, LinearArrayLayout::ColumnMajor, 1),
-        );
-        let array3 = &array1 + &array2;
-        assert_eq!(array3[[0, 0]], 8.0);
-        assert_eq!(array3[[1, 0]], 14.0);
-        assert_eq!(array3[[0, 1]], 10.0);
-        assert_eq!(array3[[1, 1]], 16.0);
-        assert_eq!(array3[[0, 2]], 12.0);
-        assert_eq!(array3[[1, 2]], 18.0);
-    }
-
-    #[test]
-    fn test_scalar_vector_ops() {
-        let a = vector![1, 2, 3];
-        let b = 2;
-
-        let c = &a + b;
-        assert_eq!(c, vector![3, 4, 5]);
-        let c = &a - b;
-        assert_eq!(c, vector![-1, 0, 1]);
-        let c = &a * b;
-        assert_eq!(c, vector![2, 4, 6]);
-        let c = &a / b;
-        assert_eq!(c, vector![0, 1, 1]);
-    }
-
-    #[test]
-    fn test_vector_vector_ops() {
-        let a = vector![1, 2, 3];
-        let b = vector![4, 5, 6];
-
-        assert_eq!(&a + &b, vector![5, 7, 9]);
-        let c = &a - &b;
-        assert_eq!(c, vector![-3, -3, -3]);
-    }
-
-    #[test]
-    #[should_panic]
-    fn test_vector_vector_ops_different_lengths_add() {
-        let a = vector![1, 2, 3];
-        let b = vector![4, 5];
-        let _ = &a + &b;
-    }
-
-    #[test]
-    #[should_panic]
-    fn test_vector_vector_ops_different_lengths_sub() {
-        let a = vector![1, 2, 3];
-        let b = vector![4, 5];
-        let _ = &a - &b;
     }
 }
